third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/grid-auto-columns-rows-expected.txt

This is a testharness.js-based test.
Found 6 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Can set 'grid-auto-columns' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'grid-auto-columns' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Setting 'grid-auto-columns' to a flexible length: -3.14fr throws TypeError
  assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
[FAIL] Can set 'grid-auto-rows' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'grid-auto-rows' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Setting 'grid-auto-rows' to a flexible length: -3.14fr throws TypeError
  assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
Harness: the test ran to completion.
